+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Added
+- Add confirmation dialog before cancelling a scheduled task
+
+### Fixed
+- Fix bug where resetting a session caused scheduled tasks and countdown tasks to be cancelled
+
 ## [0.0.27] - 2026-04-20
 
 ### Fixed